refactor(agent): unix socket adapter in slurmrestd
