+++ slurmweb/views/agent.py
@@ -13,7 +13,7 @@ from rfl.web.tokens import rbac_action, check_jwt
 
 from ..version import get_version
 from ..errors import SlurmwebCacheError, SlurmwebRestdError
-from . import SlurmrestdUnixAdapter
+from ..slurmrestd.unix import SlurmrestdUnixAdapter
 
 # Tuple used for comparaison with Slurm version retrieved from slurmrestd and
 # check for minimal supported version.
